[FEATURE][DD-413] custom deserializer for Transformation
the default implementation was confused by the 'empty' reference components used in input_components and output_components. It would have taken one of these 'empty' components to be used as the component-to-be-persisted.
This deserializer ignores these (in/out)put_components, it just collects the IDs for them. After having parsed all of the actual components, it will replace the referenced ones, optionally going to the database to lookup components, that were just referenced but were persisted before